+++ ruby/examples/app/controllers/user_registration_controller.rb
@@ -4,8 +4,8 @@ class UserRegistrationController < ApplicationController
   end
 
   def create
-    request = RegisterUser::Request.new(*user_params)
-    response = RegisterUser.new.register(request)
+    request = Users::UseCase::RegisterUser::Request.new(*user_params)
+    response = Users::UseCase::RegisterUser.new.register(request)
     @user = UserRegistration.new(request, response)
   end
 
@@ -18,8 +18,8 @@ class UserRegistrationController < ApplicationController
   class UserRegistration
     include ActiveModel::Model
 
-    attr_reader *RegisterUser::Request.members
-    attr_reader *RegisterUser::Response.members
+    attr_reader *Users::UseCase::RegisterUser::Request.members
+    attr_reader *Users::UseCase::RegisterUser::Response.members
 
     def initialize(request = nil, response = nil)
       @email = request.try(:email)
