+++ ruby/examples/lib/orders/domain/discountable_order.rb
@@ -1,0 +1,9 @@
+module Orders
+  module Domain
+    class DiscountableOrder < Struct.new(:discount)
+      def has_discount?
+        discount > 0
+      end
+    end
+  end
+end
