+++ ruby/examples/lib/users/gateway/user.rb
@@ -1,0 +1,10 @@
+module Users
+  module Gateway
+    class User
+      def create_user(attributes)
+        order = Spree::User.create(attributes)
+        order.errors
+      end
+    end
+  end
+end
